fixing display of lookbook in home and delete products relationship

+++ app/models/lookbook.rb
@@ -15,7 +15,7 @@ class Lookbook < ActiveRecord::Base
   private
 
   def update_products
-    products.delete_all
+    products.delete_all unless product_list.nil?
     selected_products = product_list.nil? ? [] : product_list.keys.collect{|id| Product.find_by_id(id)}
     selected_products.each {|os| self.products << os}
     lb = self.lookbooks_products
